minor edits
